Store crypto amounts in database as a decimal instead of a float
* Floating point arithmetic errors were leading to an inaccurate crypto balance
* In terms of cryptocurrency, this error can attribute to a significant amount of crypto
* https://0.30000000000000004.com/

+++ python-sql-scripts/addCustomers.py
@@ -73,7 +73,7 @@ create_cryptoholdings_table_sql = '''
 CREATE TABLE CryptoHoldings (
   CustomerID varchar(255),
   CryptoName varchar(255),
-  CryptoAmount float
+  CryptoAmount decimal(30,18)
 );
 '''
 cursor.execute(create_cryptoholdings_table_sql)
@@ -86,7 +86,7 @@ CREATE TABLE CryptoHistory (
   Timestamp DATETIME,
   Action varchar(255) CHECK (Action IN ('Buy', 'Sell')),
   CryptoName varchar(255),
-  CryptoAmount float
+  CryptoAmount decimal(30,18)
 );
 '''
 cursor.execute(create_cryptohistory_table_sql)
